# helpers: Remove style and script tags in plainify output