--- a/helpers/content.go
+++ b/helpers/content.go
@@ -22,7 +22,6 @@ import (
 	"fmt"
 	"html/template"
 	"os/exec"
-	"regexp"
 	"unicode"
 	"unicode/utf8"
 
@@ -188,12 +187,12 @@ func StripHTML(s string) string {
 		return s
 	}
 	s = stripHTMLReplacer.Replace(s)
-	s = stripHTMLContainers(s)
 
 	// Walk through the string removing all tags
 	b := bp.GetBuffer()
 	defer bp.PutBuffer(b)
-	var inTag, isSpace, wasSpace bool
+	var inTag, isSpace, wasSpace, ignoreTagContent bool
+	var currTag string
 	for _, r := range s {
 		if !inTag {
 			isSpace = false
@@ -201,6 +200,7 @@ func StripHTML(s string) string {
 
 		switch {
 		case r == '<':
+			currTag = ""
 			inTag = true
 		case r == '>':
 			inTag = false
@@ -208,7 +208,14 @@ func StripHTML(s string) string {
 			isSpace = true
 			fallthrough
 		default:
-			if !inTag && (!isSpace || (isSpace && !wasSpace)) {
+			if inTag {
+				currTag += string(r)
+				if currTag == "script" || currTag == "style" {
+					ignoreTagContent = true
+				} else if currTag == "/script" || currTag == "/style" {
+					ignoreTagContent = false
+				}
+			} else if !inTag && (!isSpace || (isSpace && !wasSpace)) && !ignoreTagContent {
 				b.WriteRune(r)
 			}
 		}
@@ -222,22 +229,6 @@ func StripHTML(s string) string {
 // stripEmptyNav strips out empty <nav> tags from content.
 func stripEmptyNav(in []byte) []byte {
 	return bytes.Replace(in, []byte("<nav>\n</nav>\n\n"), []byte(``), -1)
-}
-
-// stripHTMLContainers strips out script and style tags including
-// content contained within the element.
-func stripHTMLContainers(s string) string {
-	var tagsToStrip = []string{"script", "style"}
-
-	// Build a regular expression from tagsToStrip
-	var patternsToStrip = make([]string, len(tagsToStrip))
-	for i, tag := range tagsToStrip {
-		patternsToStrip[i] = "<" + tag + ">(.+?)</" + tag + ">"
-	}
-	rePattern := strings.Join(patternsToStrip, "|")
-
-	stripHTMLRegexp := regexp.MustCompile(rePattern)
-	return stripHTMLRegexp.ReplaceAllLiteralString(s, "")
 }
 
 // BytesToHTML converts bytes to type template.HTML.