Add new stripHTMLContainers function that strips out specific HTML tags including the contents contained within the element.

Fixes #3235

I originally wanted to see if I could implement this logic while walking through the string, but it seemed excessively complex to detect specific tags while walking through each codepoint in the string.  I ended up with this implementation because it felt easier to reason about.