@bep 2017-10-15
There is a benchmark written for this. If you run this before/after:

```bash
▶ benchcmp pre.bench post.bench
benchmark                old ns/op     new ns/op     delta
BenchmarkStripHTML-4     3734          38027         +918.40%

benchmark                old allocs     new allocs     delta
BenchmarkStripHTML-4     4              74             +1750.00%

benchmark                old bytes     new bytes     delta
BenchmarkStripHTML-4     736           50808         +6803.26%
```