@uncompiled 2017-10-15
That is helpful!  Let me see if I can re-implement this in some way that isn't painfully slower.  😢