@uncompiled 2017-10-16
There was a big improvement in the benchmark when [walking through the string](https://github.com/gohugoio/hugo/pull/3973/commits/eb6e5b07afc6e4c3c8654d20fae68ead1f27b4a3):
```
▶ go test -bench=BenchmarkStripHTML ./helpers
BenchmarkStripHTML-4   	  100000	     17537 ns/op
ok  	github.com/gohugoio/hugo/helpers	2.223s
```

However, I found the biggest improvement by only doing the extra work if `<style>` or `<script>` tags are present in the input string.

The benchmark now seems to be in the same ballpark using either method:

```
▶ benchcmp pre.bench post.bench 
benchmark                old ns/op     new ns/op     delta
BenchmarkStripHTML-4     3854          4274          +10.90%
```